EmergencyManagement/requirements.txt
brotli
Flask==3.0.3
GPUtil==1.4.0
GitPython==3.1.43
Jinja2
MarkupSafe
PyAudio==0.2.14
PyPDF2==3.0.1
PySocks
PyYAML
Pygments==2.18.0
SQLAlchemy==2.0.36
Send2Trash==1.8.3
SpeechRecognition==3.14.1
wave==0.0.2
Werkzeug==3.0.6
XlsxWriter==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
alpha_vantage==3.0.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi-bindings==21.2.0
argon2-cffi==23.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.0
certifi
cffi
chardet==5.2.0
charset-normalizer
click>=8.1.8,<8.2
cmdstanpy==1.2.5
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash-vtk==0.0.9
dash==2.18.2
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
docstring_parser==0.16
duckduckgo_search==8.0.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.12
fastjsonschema==2.21.1
filelock
fonttools==4.55.2
fpdf2==2.8.1
fpdf==1.7.2
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
graphviz==0.20.1
greenlet==3.1.1
h11==0.14.0
h2
holidays==0.68
hpack
httpcore==1.0.7
httpx-sse==0.4.0
httpx==0.28.0
huggingface-hub==0.26.3
hyperframe
idna
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.34.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema-specifications==2024.10.1
jsonschema==4.23.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter==1.1.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langchain-community==0.0.38
langchain-core==0.1.53
langchain-text-splitters==0.0.2
langchain==0.1.11
langsmith==0.1.147
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==5.3.0
markdown-it-py==3.0.0
marshmallow==3.23.1
matplotlib-inline==0.1.7
matplotlib==3.10.1
mdurl==0.1.2
mistune==3.1.2
mpmath
msgpack==1.1.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.15.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.2
numpy
ollama==0.4.7
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.12
overrides==7.7.0
packaging==23.2
pandas-datareader==0.10.0
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
peewee==3.17.9
pillow
platformdirs==4.3.6
plotly-geo==1.0.0
plotly==5.24.1
pooch==1.8.2
primp==0.15.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
prophet==1.1.6
protobuf==4.25.6
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pycparser
pydantic-settings==2.6.1
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pydot==1.4.2
pygame==2.6.1
pyparsing==3.2.1
pypdf==5.1.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-pptx==0.6.21
pyttsx3==2.98
pytz==2024.2
pyzmq==26.2.1
referencing==0.35.1
regex==2024.11.6
reportlab==4.2.5
requests
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
scooby==0.10.0
seaborn==0.13.2
segment-anything==1.0
sentence-transformers==3.3.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
sqlparse==0.4.4
stack-data==0.6.3
stanio==0.5.1
starlette==0.46.2
statsmodels==0.14.4
streamlit==1.32.0
sympy==1.13.1
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
typer==0.15.4
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions
tzdata==2024.2
ultralytics-thop==2.0.14
ultralytics==8.3.146
uri-template==1.3.0
urllib3
vtk>=9.5.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
yarl==1.18.3
yfinance==0.2.54
zipp==3.21.0
zstandard==0.23.0
agno==1.5.1
